(wip) fix(oss): fix LOG/fmt issues for "stock" fmt
Summary:
Several build issues because sapling currently depends on the custom
behavior of an internal fork of the C++ fmt library.

Test Plan:

Confirm the following gets past the compilation step without formatter
issues
```
act --container-architecture linux/x86_64 -W ".github/workflows/edenfs_linux.yml"
```

+++ eden/fs/inodes/test/CheckoutTest.cpp
@@ -691,7 +691,7 @@ void runModifyConflictTests(CheckoutMode checkoutMode) {
   for (StringPiece path : {"a/b/aaa.txt", "a/b/mmm.txt", "a/b/zzz.tzt"}) {
     for (auto loadType : kAllLoadTypes) {
       SCOPED_TRACE(fmt::format(
-          "path {} load type {} force={}", path, loadType, checkoutMode));
+          "path {} load type {} force={}", path, loadType, static_cast<int>(checkoutMode)));
       testModifyConflict(
           path,
           loadType,
